Compiles with GCC 4. Fix hidden toolbar. Fix LIST command adding trailing space. Beep on transfer error and show last output. Hide password in prompt dialog. Clarify SSH authentication errors. Fix new profile having ask password option set. Add wspiapi.h: allow running on w2k. Ask to accept new host key instead of always rejecting. Fix master key of 8 characters being stored as 7. Session reuse for FTP(E)S (untested). Version now 0.2.4 .
git-svn-id: svn://svn.code.sf.net/p/nppftp/code/NppFTP@14 e1eea60f-96cb-45df-9f85-ae3a5fd22bdd

+++ UTCP/source/UTDataSource.cpp
@@ -811,7 +811,7 @@ int	CUT_MapFileDataSource::Open(OpenMsgType type)
 		m_lnActualSize.QuadPart = m_lnSize.QuadPart;
 
 	// Initialize increment value. Size divided by 8 ( or shifted by 3)
-	m_lnIncrement.QuadPart = max(Int64ShraMod32(m_lnSize.QuadPart, 3), 4096);
+	m_lnIncrement.QuadPart = max(Int64ShraMod32(m_lnSize.QuadPart, 3), LONGLONG(4096));
 
 	// If file is opened in append mode - increase it size
 	if(type == UTM_OM_APPEND || (type == UTM_OM_WRITING && m_lnSize.HighPart == 0 && m_lnSize.LowPart == 0 ))
